Move utilities to the base class
