+++ sample-app/src/test/scala/example/readmodel/rdb/DatabaseConcertRepositorySpecBase.scala
@@ -7,6 +7,11 @@ import org.scalatest.time.{ Milliseconds, Span }
 import slick.basic.DatabaseConfig
 import slick.jdbc.JdbcProfile
 
+import java.sql.Timestamp
+import java.time.ZonedDateTime
+import java.time.temporal.ChronoUnit
+import scala.concurrent.duration.FiniteDuration
+
 abstract class DatabaseConcertRepositorySpecBase() extends ActorSpecBase() with BeforeAndAfter {
   // NOTE: すべての操作が H2DB 互換であり、H2DB でテストにパスすればよいと妥協している
   protected val config: Config = ConfigFactory.parseString("""
@@ -44,6 +49,27 @@ abstract class DatabaseConcertRepositorySpecBase() extends ActorSpecBase() with 
     import databaseService.profile.api._
     val schema = concerts.schema ++ updaterOffsets.schema
     databaseService.database.run(schema.truncate).futureValue
+  }
+
+  // 現在時刻を秒精度で返す。
+  // 秒未満はデータベースによってデフォルト値やサポート範囲が異なるため秒精度に丸める。
+  // https://mariadb.com/kb/en/timestamp/#supported-values
+  // https://www.h2database.com/html/datatypes.html#timestamp_type
+  protected def nowInSeconds: ZonedDateTime = {
+    ZonedDateTime.now.truncatedTo(ChronoUnit.SECONDS)
+  }
+
+}
+
+object DatabaseConcertRepositorySpecBase {
+
+  implicit final class RichZonedDateTime(val time: ZonedDateTime) extends AnyVal {
+    def +(duration: FiniteDuration): ZonedDateTime = {
+      time.plusNanos(duration.toNanos)
+    }
+    def toSQLTimestamp: Timestamp = {
+      Timestamp.from(time.toInstant)
+    }
   }
 
 }
